### 2020-03-05 #374 @CamiloCarneiro
eneRgetics
Submitting Author: Name (@CamiloCarneiro)
Repository:   https://github.com/CamiloCarneiro/eneRgetics

---

```
Package: eneRgetics
Title: Estimates the Maintenance Metabolism of shorebirds
Version: 0.0.0.9000
Description: eneRgetics provides a fast and straightforward way to calculate the
    Maintenance Metabolism of shorebirds. It has functions to retrieve the necessary
    environmental data to calculate it, to retrieve tidal data and to create
    categories for the tide state in relation to low tide. Additionally, it provides
    a function to calculate the Basal Metabolic Rate of shorebirds.
    checkmate (>= 1.8.5),
    camsRad (>= 0.3.0),
    darksky (>= 1.3.0),
    lubridate (>= 1.1.7),
    RCurl (>= 1.95-4.8),
    jsonlite(>= 1.5)
VignetteBuilder: rmarkdown
```

- Please indicate which category or categories from our [package fit policies](https://ropensci.github.io/dev_guide/policies.html#package-categories) this package falls under: (Please check an appropriate box below.:

The main advantage of this package is to easily retrieve, from online sources, the data required to calculate the maintenance metabolism of shorebirds, such as weather and tidal data.

Researchers interested in determining the maintenance metabolism of shorebirds. For example, to compare the maintenance metabolism across a wide wintering range (which most shorebirds have) allows the investigation of costs/benefits and their trade-off with certain life-history traits (e.g. migration distance). This package creates a fast route to gather the necessary data for such analysis. These kind of analysis are relevant when we think about the global changes in the climate, that might lead to a change in the energetic requirements of animals.

We are not aware of any other package that aims to accomplish the same thing.

This is the first package we are preparing.
