### 2020-03-17 #374 @annakrystalli
Dear @CamiloCarneiro,

Thanks again for your submission. We have been discussing whether the package is in scope and have concluded that unfortunately it is not for the following reasons:

- The package has been submitted under data retrieval
- The main innovation as we see it is retrieval of the tide data. However, the world tides API is a paid service and calls to it are costed. Users seem to get 100 free credits (calls) but then need to be paid for https://www.worldtides.info/developer.
- The weather is spatially limited and just combines data sourced through packages `darksky` and `camsRad`) and then combine the data. While useful, this does not add significant value.
- The main function of the package is the maintenance metabolism calculation `mmaint`, which as useful as it might be to domain specialists, renders it ultimately out of scope under data retrieval.

Having said that we still think your package may well be of use to domain specialist so we encourage you to continue making it available to your community should you wish to.

#### Some general advice for improving the package:

- It is not clear from the README that the tide data is sourced from a paid service. We advise you make details of potential costs much clearer in the README.
- We also recommend you mention the various APIs users need to get tokens for in the README (perhaps following installation instructions) and provide more details on how to get each API token and how to handle them in a separate vignette. API tokens need to be handled securely and you may need to educate users on how to do so. See for example instructions on how to [handle `GITHUB_PAT`](https://happygitwithr.com/github-pat.html) API tokens.
- Your `build` and `coverage` badges in the README are showing **unknown**. You might want to trigger a build to ensure the data required for the badges is available.

Many thanks again for your patience as we assessed the package.
