### 2020-03-17 #374 @CamiloCarneiro
Dear @annakrystalli,

Thank you for your time to assess the package and for the advice to improve it. Much appreciated.

Best regards.
